logger: Move logger to separate object
