+++ app/src/main/java/us/spotco/malwarescanner/Database.kt
@@ -26,7 +26,7 @@ import com.google.common.hash.BloomFilter
 import com.google.common.hash.Funnels
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
-import us.spotco.malwarescanner.HypatiaLogger.Companion.log
+import us.spotco.malwarescanner.HypatiaLogger.log
 import us.spotco.malwarescanner.Utils.getDatabaseURL
 import us.spotco.malwarescanner.Utils.getSigningKey
 import us.spotco.malwarescanner.Utils.threadPoolExecutor
