+++ app/src/main/java/us/spotco/malwarescanner/Hypatia.kt
@@ -33,7 +33,7 @@ class Hypatia : Application() {
     fun setMainActivity(activity: MainActivity?) {
         mainActivityRef = WeakReference<MainActivity?>(activity)
         if (activity != null) {
-        HypatiaLogger.instance?.setLogView(activity.getLogView())
+            HypatiaLogger.setLogView(activity.getLogView())
         } else {
             Log.e("Hypatia", "COULDN'T START HYPATIA LOGGER")
         }
