+++ app/src/main/java/us/spotco/malwarescanner/HypatiaLogger.kt
@@ -44,12 +44,7 @@ private const val TAG = "Hypatia Logger"
  * TODO use a separate log page to list this logs, not MainActivity
  *
  */
-class HypatiaLogger
-    /*
-     * Utilities to ensure Logger working correctly and getting logs from
-     * multiple sources
-     */
-private constructor() {
+object HypatiaLogger {
     private val logQueue = ConcurrentLinkedQueue<String?>()
     private var logViewRef: WeakReference<TextView>? = null
     private val uiHandler = Handler(Looper.getMainLooper())
@@ -81,87 +76,70 @@ private constructor() {
         }
     }
 
-    companion object {
-        @JvmStatic
-        var instance: HypatiaLogger? = null
-            get() {
-                if (field == null) {
-                    synchronized(HypatiaLogger::class.java) {
-                        if (field == null) {
-                            field = HypatiaLogger()
-                        }
-                    }
-                }
-                return field
+    /**
+     * Logs process outputs to user interface
+     * prints out also verbose/debug if app is debug version
+     *
+     * @param content log content
+     * @param type    log type
+     */
+    /**
+     * Overloader of actual `log()` function
+     *
+     * @see .log
+     * @param content log content
+     */
+    internal fun log(content: String, type: LogType = LogType.DEFAULT) {
+        var message = content
+        var userFacing = true
+
+        // toggle these to show/hide DEBUG/VERBOSE and VERY_VERBOSE in debug app builds
+        val ENABLE_DEBUG = false
+        val ACTIVATE_VERY_VERBOSE = false
+        when (type) {
+            LogType.VERY_VERBOSE -> if (ACTIVATE_VERY_VERBOSE) {
+                Log.v("Hypatia Logger", message)
             }
-            private set
-        /**
-         * Logs process outputs to user interface
-         * prints out also verbose/debug if app is debug version
-         *
-         * @param content log content
-         * @param type    log type
-         */
-        /**
-         * Overloader of actual `log()` function
-         *
-         * @see .log
-         * @param content log content
-         */
-        @JvmStatic
-        @JvmOverloads
-        internal fun log(content: String, type: LogType = LogType.DEFAULT) {
-            var content = content
-            var userFacing = true
 
-            // toggle these to show/hide DEBUG/VERBOSE and VERY_VERBOSE in debug app builds
-            val ENABLE_DEBUG = false
-            val ACTIVATE_VERY_VERBOSE = false
-            when (type) {
-                LogType.VERY_VERBOSE -> if (ACTIVATE_VERY_VERBOSE) {
-                    Log.v("Hypatia Logger", content)
+            LogType.VERBOSE -> {
+                Log.v("Hypatia Logger", message)
+                message = "VERBOSE: $message"
+                if (!BuildConfig.DEBUG || !ENABLE_DEBUG) {
+                    userFacing = false
                 }
-
-                LogType.VERBOSE -> {
-                    Log.v("Hypatia Logger", content)
-                    content = "VERBOSE: $content"
-                    if (!BuildConfig.DEBUG || !ENABLE_DEBUG) {
-                        userFacing = false
-                    }
-                }
-
-                LogType.DEBUG -> {
-                    Log.d("Hypatia Logger", content)
-                    content = "DEBUG: $content"
-                    if (!BuildConfig.DEBUG || !ENABLE_DEBUG) {
-                        userFacing = false
-                    }
-                }
-
-                LogType.INFO -> {
-                    Log.i("Hypatia Logger", content)
-                    content = "INFO: $content"
-                }
-
-                LogType.WARNING -> {
-                    Log.w("Hypatia Logger", content)
-                    content = "WARNING: $content"
-                }
-
-                LogType.ERROR -> {
-                    Log.e("Hypatia Logger", content)
-                    content = "ERROR: $content"
-                }
-
-                else -> Log.d("Hypatia Logger", content)
             }
-            try {
-                if (userFacing && (type != LogType.VERY_VERBOSE)) {
-                    instance!!.logUtil(content + (if (type == LogType.NO_NEWLINE) "" else "\n"))
+
+            LogType.DEBUG -> {
+                Log.d("Hypatia Logger", message)
+                message = "DEBUG: $message"
+                if (!BuildConfig.DEBUG || !ENABLE_DEBUG) {
+                    userFacing = false
                 }
-            } catch (e: Exception) {
-                Log.d(TAG, "Logger error: ", e)
             }
+
+            LogType.INFO -> {
+                Log.i("Hypatia Logger", message)
+                message = "INFO: $message"
+            }
+
+            LogType.WARNING -> {
+                Log.w("Hypatia Logger", message)
+                message = "WARNING: $message"
+            }
+
+            LogType.ERROR -> {
+                Log.e("Hypatia Logger", message)
+                message = "ERROR: $message"
+            }
+
+            else -> Log.d("Hypatia Logger", message)
+        }
+        try {
+            if (userFacing && (type != LogType.VERY_VERBOSE)) {
+                logUtil(message + (if (type == LogType.NO_NEWLINE) "" else "\n"))
+            }
+        } catch (e: Exception) {
+            Log.d(TAG, "Logger error: ", e)
         }
     }
 }
