+++ app/src/main/java/us/spotco/malwarescanner/LinkScannerService.kt
@@ -28,7 +28,7 @@ import android.view.accessibility.AccessibilityEvent
 import android.view.accessibility.AccessibilityNodeInfo
 import android.widget.Toast
 import androidx.core.app.NotificationCompat
-import us.spotco.malwarescanner.HypatiaLogger.Companion.log
+import us.spotco.malwarescanner.HypatiaLogger.log
 import java.util.Locale
 import java.util.Random
 import java.util.concurrent.ConcurrentSkipListMap
