+++ app/src/main/java/us/spotco/malwarescanner/MainActivity.kt
@@ -51,8 +51,8 @@ import kotlinx.coroutines.launch
 import us.spotco.malwarescanner.DatabaseSource.MT_MIRROR
 import us.spotco.malwarescanner.Hypatia.Companion.appContext
 import us.spotco.malwarescanner.Hypatia.Companion.getInstance
-import us.spotco.malwarescanner.HypatiaLogger.Companion.log
-import us.spotco.malwarescanner.HypatiaLogger.Companion.instance
+import us.spotco.malwarescanner.HypatiaLogger.log
+import us.spotco.malwarescanner.HypatiaLogger.setLogView
 import java.io.File
 import java.util.concurrent.ConcurrentLinkedQueue
 
@@ -85,7 +85,7 @@ class MainActivity : AppCompatActivity() {
         logView!!.movementMethod = ScrollingMovementMethod()
         logView!!.setTextIsSelectable(true)
 
-        instance!!.setLogView(logView)
+        setLogView(logView)
 
         log(getString(R.string.app_disclaimer) + "\n\n\n")
         log(getString(R.string.app_copyright))
