+++ app/src/main/java/us/spotco/malwarescanner/MalwareScanner.kt
@@ -31,7 +31,7 @@ import androidx.core.app.NotificationCompat
 import com.google.common.hash.BloomFilter
 import kotlinx.coroutines.runBlocking
 import kotlinx.coroutines.withTimeoutOrNull
-import us.spotco.malwarescanner.HypatiaLogger.Companion.log
+import us.spotco.malwarescanner.HypatiaLogger.log
 import java.io.File
 import java.io.FileInputStream
 import java.io.InputStream
