+++ app/src/main/java/us/spotco/malwarescanner/NotificationPromptActivity.kt
@@ -26,7 +26,7 @@ import android.content.pm.PackageManager
 import android.os.Bundle
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
-import us.spotco.malwarescanner.HypatiaLogger.Companion.log
+import us.spotco.malwarescanner.HypatiaLogger.log
 import java.io.File
 import java.util.Objects
 import androidx.core.net.toUri
