+++ app/src/main/java/us/spotco/malwarescanner/ShareScanner.kt
@@ -28,7 +28,7 @@ import android.provider.OpenableColumns
 import android.util.Log
 import androidx.appcompat.app.AppCompatActivity
 import us.spotco.malwarescanner.Hypatia.Companion.appContext
-import us.spotco.malwarescanner.HypatiaLogger.Companion.log
+import us.spotco.malwarescanner.HypatiaLogger.log
 import java.io.File
 import java.io.FileOutputStream
